Add new ini parser; merges commonly named sections--allows #INCLUDE to be more effective
This patch adds a ini-merging ability whereby tag=value items are merged
together into the same effective section.  Consider the following ini file:
----begin----
[A]
var0 = 0

[B]
var0 = 12

[A]
var0 = 123
var1 = hi
---end----

With the old parser, only the first instance of the 'A' section would be
recognized.  The second instance of the 'A' section would be entirely
ignored.  With this new parser, the final 'A' section contains two values
for var0 and one value for var1.

The major benefit of this patch is that the #INCLUDE command can be used
much more effectively.  As an example, one could define a configuration
file that is fairly generic and _almost_ fits perfectly for a whole series
of machines.  Then, in order to specialize the configuration for a
particular machine, one can simply set the specialized tag=value items
first in the specialized ini-file; following these specialized tag=value
items, one simply uses #INCLUDE to import the generic configuration, were
the specialized tag=value items are first taken from the specialized ini
file.

One additional feature is added with this patch:
If one uses the ":=" operator, instead of the simple "=" operator, one can
override all previous values for that particular [SECTION]VARIABLE.  In
other words, before using a ":=" operator, the above example yields two
separate values for [A]var0.  If a subsequent assignment to [A]var0 looked
like:
-----beg-snip----
[A]
var0 := 1010
-----end-snip----
then the previous values for [A]var0 would be discarded and only this last
value remains.

Note:  The new ':=' override operator is implemented both in the c++ _and_
tcl code.

Signed-off-by: Spencer E Olson <olsonse@umich.edu>